Update test packages
The project was using a very old version of Mocah and should packages.
This brings them up to the current stable versions of each.  This
actually exposed a mistake in the tests as written, which forgot to
properly close the Opened Captions server before declaring the tests
over in some cases.

The tests still have an issue, which also exposes a mistake that will
need to be fixed in a separate commit.

Issue #15 Modernize development environment