Add stream cleaning to close
Opened Captions has a close method which, in theory, should stop its
operations.  It turns out it actually didn't, and in the case of the
random stream the interval kept going even after a close command was
issued.

This fixes this oversight, and also means that the unit tests properly
end.  The reason this got noticed is that about a year ago Mocha decided
to no longer automatically terminate the test process.  You can read
more about that https://boneskull.com/mocha-v4-nears-release/#mochawontforceexit