Disconnect socket listener in tests
Another exposed mistake thanks to Mocha's change to process killing
(I'm glad they made that change to be honest...)

The tests which created a SocketIO client to test opened captions were
passing, and then failing because after they ended the opened captions
server closed but the client was not disconnected first.  This change
disconnects the client before turning off the server.