Remove use of lodash
Lodash is a library that simply is not needed any more, and it had
a major security vulnerability to boot.  This swaps out some of what
lodash was offering since much of it is now baked into ES2015.  Lodash
itself will be removed in a future commit

Issue #15 Modernize development environment